+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change log
 
-## 0.14.0 (2024-09-24)
+## 0.15.0 (2024-09-30)
+
+### Features
+
+- core: Save release visited history
+- popup: Show releases history sorted by visited date by default
+- popup: Show latest visited releases on the Dashboard tab
+
+## 0.14.1 (2024-09-24)
 
 ### Fixed